requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
langchain-ollama
sentence-transformers
chromadb
ollama
pytesseract
torch
torchvision
numpy
opencv-python
Pillow
rouge-score
jiwer
scikit-image
scikit-learn